+++ src/control_plane/shared_infrastructure/mongodb/mod.rs
@@ -1,17 +1,27 @@
 pub(crate) use mongodb_logical_resource_plan::MongoDbLogicalResourcePlan;
 pub(crate) use mongodb_plan_error::MongoDbPlanError;
+pub(crate) use mongodb_preparation_error::MongoDbPreparationError;
+pub(crate) use mongodb_preparation_options::MongoDbPreparationOptions;
 pub(crate) use mongodb_project_resources::MongoDbProjectResources;
 pub(crate) use mongodb_shared_instance_plan::MongoDbSharedInstancePlan;
 pub(crate) use mongodb_shared_instance_plan_options::MongoDbSharedInstancePlanOptions;
 pub(crate) use plan_mongodb_project_resources::plan_mongodb_project_resources;
+pub(crate) use prepare_mongodb_shared_instances::prepare_mongodb_shared_instances;
+pub(crate) use prepared_mongodb_shared_instance::PreparedMongoDbSharedInstance;
 pub(crate) use provision_mongodb_logical_resource::provision_mongodb_logical_resource;
 pub(crate) use reconcile_mongodb_project_resources::reconcile_mongodb_project_resources;
+pub(crate) use reconcile_prepared_mongodb_instance::reconcile_prepared_mongodb_instance;
 
 mod mongodb_logical_resource_plan;
 mod mongodb_plan_error;
+mod mongodb_preparation_error;
+mod mongodb_preparation_options;
 mod mongodb_project_resources;
 mod mongodb_shared_instance_plan;
 mod mongodb_shared_instance_plan_options;
 mod plan_mongodb_project_resources;
+mod prepare_mongodb_shared_instances;
+mod prepared_mongodb_shared_instance;
 mod provision_mongodb_logical_resource;
 mod reconcile_mongodb_project_resources;
+mod reconcile_prepared_mongodb_instance;